+++ AnnotationDialog/ImagePreviewWidget.cpp
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2009-2010 Hassan Ibraheem <hasan.ibraheem@gmail.com>
 // SPDX-FileCopyrightText: 2010 Tuomas Suutari <tuomas@nepnep.net>
 // SPDX-FileCopyrightText: 2010-2022 Jesper K. Pedersen <jesper.pedersen@kdab.com>
-// SPDX-FileCopyrightText: 2013-2023 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2013-2024 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
 // SPDX-FileCopyrightText: 2014-2019 Tobias Leupold <tl@stonemx.de>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
@@ -159,6 +159,13 @@ void ImagePreviewWidget::configure(QList<DB::ImageInfo> *imageList, bool singleE
     m_copyPreviousBut->setEnabled(m_singleEdit);
     m_rotateLeft->setEnabled(m_singleEdit);
     m_rotateRight->setEnabled(m_singleEdit);
+}
+
+void ImagePreviewWidget::reconfigure(QList<DB::ImageInfo> *imageList, int index)
+{
+    m_imageList = imageList;
+    m_current = index;
+    setImage(m_imageList->at(m_current));
 }
 
 void ImagePreviewWidget::slotPrev()